+++ ff.h
@@ -101,7 +101,8 @@ Boolean readparams(Handle h,Boolean alerts,char **reason);
 void convert_premiere_to_photoshop(PARM_T* photoshop, PARM_T_PREMIERE* premiere);
 Boolean read8bfplugin(StandardFileReply *sfr,char **reason);
 Handle readfileintohandle(FILEREF r);
-Boolean readfile(StandardFileReply *sfr,char **reason);
+Boolean readfile_afs_pff(StandardFileReply* sfr, char** reason);
+Boolean readfile_ffx(StandardFileReply* sfr, char** reason);
 Boolean readPARM(Ptr h,PARM_T *parm,char **reasonstr,int fromwin);
 
 // from save.c