+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Tab-Order has been changed (*)
 - Fixed incompatibility with older versions of Wine (Windows emulator for Linux systems)
 - In the Make dialog, if you clicked "OK" and then cancelled the save dialog, you didn't get back to the Make-dialog. Fixed.
+- "Filters Unlimited" filters (FFX files) can now be loaded. However, only the basic Filter Factory commands are supported.
 
 (*) This bug/solution was tested on Windows but needs to be verified and/or implemented on Mac.